:recycle: refactor(1d) move activity answers files
