+++ api/lib/domain/models/ValidatorAlwaysOK.js
@@ -1,6 +1,6 @@
 import { Validation } from './Validation.js';
 import { Validator } from './Validator.js';
-import { AnswerStatus } from './AnswerStatus.js';
+import { AnswerStatus } from '../../../src/school/domain/models/AnswerStatus.js';
 
 class ValidatorAlwaysOK extends Validator {
   assess() {
