+++ api/tests/integration/infrastructure/repositories/certification-assessment-repository_test.js
@@ -3,7 +3,7 @@ import { NotFoundError } from '../../../../lib/domain/errors.js';
 import * as certificationAssessmentRepository from '../../../../lib/infrastructure/repositories/certification-assessment-repository.js';
 import { CertificationAssessment } from '../../../../lib/domain/models/CertificationAssessment.js';
 import { Challenge } from '../../../../src/shared/domain/models/Challenge.js';
-import { AnswerStatus } from '../../../../lib/domain/models/AnswerStatus.js';
+import { AnswerStatus } from '../../../../src/school/domain/models/AnswerStatus.js';
 import _ from 'lodash';
 
 describe('Integration | Infrastructure | Repositories | certification-assessment-repository', function () {
