+++ api/tests/school/integration/domain/usecases/get-next-challenge_test.js
@@ -4,7 +4,7 @@ import { Activity } from '../../../../../src/school/domain/models/Activity.js';
 import * as activityRepository from '../../../../../src/school/infrastructure/repositories/activity-repository.js';
 import * as assessmentRepository from '../../../../../src/shared/infrastructure/repositories/assessment-repository.js';
 import * as challengeRepository from '../../../../../src/school/infrastructure/repositories/challenge-repository.js';
-import * as activityAnswerRepository from '../../../../../lib/infrastructure/repositories/activity-answer-repository.js';
+import * as activityAnswerRepository from '../../../../../src/school/infrastructure/repositories/activity-answer-repository.js';
 import * as missionAssessmentRepository from '../../../../../src/school/infrastructure/repositories/mission-assessment-repository.js';
 import { getNextChallenge } from '../../../../../src/school/domain/usecases/get-next-challenge.js';
 import * as learningContentBuilder from '../../../../tooling/learning-content-builder/index.js';
