+++ api/tests/school/integration/infrastructure/repositories/activity-answer-repository_test.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
-import { databaseBuilder, domainBuilder, expect, knex } from '../../../test-helper.js';
-import { AnswerStatus } from '../../../../lib/domain/models/AnswerStatus.js';
-import * as activityAnswerRepository from '../../../../lib/infrastructure/repositories/activity-answer-repository.js';
-import { Activity } from '../../../../src/school/domain/models/Activity.js';
+import { databaseBuilder, domainBuilder, expect, knex } from '../../../../test-helper.js';
+import { AnswerStatus } from '../../../../../src/school/domain/models/AnswerStatus.js';
+import * as activityAnswerRepository from '../../../../../src/school/infrastructure/repositories/activity-answer-repository.js';
+import { Activity } from '../../../../../src/school/domain/models/Activity.js';
 
 describe('Integration | Repository | activityAnswerRepository', function () {
   describe('#findByActivity', function () {
