+++ api/tests/tooling/domain-builder/factory/build-validation.js
@@ -1,4 +1,4 @@
-import { AnswerStatus } from '../../../../lib/domain/models/AnswerStatus.js';
+import { AnswerStatus } from '../../../../src/school/domain/models/AnswerStatus.js';
 import { Validation } from '../../../../lib/domain/models/Validation.js';
 
 const buildValidation = function ({ result = AnswerStatus.OK, resultDetails = 'Bravo' } = {}) {
