+++ api/tests/unit/domain/models/AnswerStatus_test.js
@@ -2,7 +2,7 @@ import chai from 'chai';
 
 const { expect } = chai;
 
-import { AnswerStatus } from '../../../../lib/domain/models/AnswerStatus.js';
+import { AnswerStatus } from '../../../../src/school/domain/models/AnswerStatus.js';
 
 describe('AnswerStatus', function () {
   context('AnswerStatus#isOK', function () {
