+++ api/tests/unit/domain/models/Validator_test.js
@@ -1,4 +1,4 @@
-import { AnswerStatus } from '../../../../lib/domain/models/AnswerStatus.js';
+import { AnswerStatus } from '../../../../src/school/domain/models/AnswerStatus.js';
 import { Validator } from '../../../../lib/domain/models/Validator.js';
 import { Validation } from '../../../../lib/domain/models/Validation.js';
 import { expect, domainBuilder } from '../../../test-helper.js';
