+++ api/tests/unit/domain/services/smart-random/smart-random_test.js
@@ -1,5 +1,5 @@
 import { expect, domainBuilder } from '../../../../test-helper.js';
-import { AnswerStatus } from '../../../../../lib/domain/models/AnswerStatus.js';
+import { AnswerStatus } from '../../../../../src/school/domain/models/AnswerStatus.js';
 import * as SmartRandom from '../../../../../lib/domain/services/algorithm-methods/smart-random.js';
 import _ from 'lodash';
