+++ api/tests/unit/domain/usecases/simulate-flash-deterministic-assessment-scenario_test.js
@@ -1,6 +1,6 @@
 import { domainBuilder, expect, sinon } from '../../../test-helper.js';
 import { simulateFlashDeterministicAssessmentScenario } from '../../../../src/certification/flash-certification/domain/usecases/simulate-flash-deterministic-assessment-scenario.js';
-import { AnswerStatus } from '../../../../lib/domain/models/AnswerStatus.js';
+import { AnswerStatus } from '../../../../src/school/domain/models/AnswerStatus.js';
 import { config } from '../../../../src/shared/config.js';
 import _ from 'lodash';
