+++ api/tests/unit/infrastructure/adapters/answer-status-json-api-adapter_test.js
@@ -1,5 +1,5 @@
 import { AnswerStatusJsonApiAdapter } from '../../../../lib/infrastructure/adapters/answer-status-json-api-adapter.js';
-import { AnswerStatus } from '../../../../lib/domain/models/AnswerStatus.js';
+import { AnswerStatus } from '../../../../src/school/domain/models/AnswerStatus.js';
 
 import chai from 'chai';
